:recycle: refactor: split App in Movies component
create a custom hook to fetch movies
